Move Main to another class, and call CLInterface from there
This is a temporary move to temporarily fix the workflow -- current CLInterface doesn't work with Chris's workflow and has a few general bugs. This commit makes it so I don't have to comment a bunch of stuff out every time I merge. @banks will fix the CLInterface soon enough and we'll move Main back to there.